backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Azure OpenAI
openai>=1.12.0

# PDF processing
PyMuPDF>=1.23.0

# Image processing
Pillow>=10.2.0

# Type hints
pydantic>=2.5.0

